Enhancement：Refactored code for reuse.
